+++ src/test/scala/com/atomist/rug/runtime/plans/ExampleRugFunction.scala
@@ -1,6 +1,7 @@
 package com.atomist.rug.runtime.plans
 
 import com.atomist.param._
+import com.atomist.rug.RugRuntimeException
 import com.atomist.rug.spi.Handlers.Status
 import com.atomist.rug.spi._
 
@@ -15,7 +16,11 @@ class ExampleRugFunction
   failure.setRequired(false)
   failure.setDefaultValue("false")
 
-  override def parameters = Seq(new Parameter("thingy"), failure)
+  val exception = new Parameter("exception")
+  exception.setRequired(false)
+  exception.setDefaultValue("false")
+
+  override def parameters = Seq(new Parameter("thingy"), failure, exception)
 
   override def secrets: Seq[Secret] = {
     if (ExampleRugFunction.clearSecrets) {
@@ -32,7 +37,10 @@ class ExampleRugFunction
     validateParameters(parameters)
     if(parameters.parameterValueMap.contains("fail") && parameters.parameterValueMap("fail").getValue == "true"){
       FunctionResponse(Status.Failure,Some("Something went wrong :("), Some(500), StringBodyOption(parameters.parameterValueMap("fail").getValue.toString))
-    }else{
+    }else if(parameters.parameterValueMap.contains("exception") && parameters.parameterValueMap("exception").getValue == "true"){
+      throw new RugRuntimeException(null, "uh oh")
+    }
+    else {
       FunctionResponse(Status.Success,Some("It worked! :p"), Some(204), StringBodyOption(parameters.parameterValueMap("thingy").getValue.toString))
     }
   }